refactor(launcher): extract shared launcher logic into util.ts
Consolidate duplicated detectLibc, getPackageSuffix, and findBinary
functions from cli.ts and mcp.ts into a shared scripts/launcher/util.ts
module. Each launcher is reduced from ~119 lines to ~17 lines.

Also apply minor else-block style fix in scripts/bump-version.ts.

Closes #111
